// src/enums/fmlxUIEnum.ts
export enum SizeFmlxUiEnum {
    SMALL = 'sm',
    EXTRA_SMALL = 'xs',
    MEDIUM = 'md',
    LARGE = 'lg',
    EXTRA_LARGE = 'xl',
}

export enum TypeFmlxUiEnum {
    POSITIVE = 'positive',
    PRIMARY = 'primary',
    DANGER = 'danger',
    BASIC = 'basic',
}

export enum VariantFmlxUiEnum {
    ACTION_BUTTON = 'action-button',
    CONTAIN = 'contain',
    OUTLINE = 'outline',
    PLAIN = 'plain',
}